Disconnect settings on disable/destroy
We connect to setting change events, but never disconnect.
Destroying the object doesn't disconnect the events. Disabling
and re-enabling the extension then caused repeated updates.
And locking and unlocking the screen caused the extension to
be disabled and re-enabled.

We now clean up the setting events when the object gets destroyed.